Improve undo for fragmented actions
